New homepage content + function fixes (1)
Thanks to a meeting with Matt Jansen, I figured out how to add content to the homepage. I also went in and changed some of the syntax for the SMITE functions (e.g., model m is now model x). 

However, on my sample website, it looks like these changes haven't updated. So, I'm uploading these changes to GitHub now, and we'll see if that changes how the functions work on the website.